sepalate tests to treekem and enclave_key
