+++ nodes/key-vault/src/test/enclave_key.rs
@@ -1,0 +1,21 @@
+// use crate::{handlers::handle_health_check, Server as KeyVaultServer};
+// use actix_web::{http::StatusCode, test, web, App};
+// use anonify_eth_driver::utils::*;
+// use frame_common::crypto::Ed25519ChallengeResponse;
+// use frame_config::{ANONIFY_ABI_PATH, FACTORY_ABI_PATH, PJ_ROOT_DIR};
+// use frame_host::EnclaveDir;
+// use frame_sodium::{SodiumCiphertext, SodiumPubKey};
+// use once_cell::sync::Lazy;
+// use rand_core::{CryptoRng, RngCore};
+// use serde_json::json;
+// use state_runtime_node_server::{handlers::*, Server as ERC20Server};
+// use std::{env, fs, path::Path, str::FromStr, sync::Arc};
+// use web3::{contract::Options, types::Address};
+//
+// use super::*;
+
+// TODO: unimplemented
+#[actix_rt::test]
+async fn test_enclave_key_backup() {
+
+}
